@@ -4,6 +4,7 @@
 
 package dev.fluttercommunity.plus.battery;
 
+import android.annotation.TargetApi;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.ContextWrapper;
@@ -15,6 +16,8 @@ import android.os.Build.VERSION;
 import android.os.Build.VERSION_CODES;
 import android.os.PowerManager;
 import android.provider.Settings;
+import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.EventChannel;
@@ -24,7 +27,6 @@ import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
-import io.flutter.plugin.common.PluginRegistry;
 import java.util.Locale;
 
 /** BatteryPlusPlugin */
@@ -38,12 +40,6 @@ public class BatteryPlusPlugin implements MethodCallHandler, StreamHandler, Flut
   public static final String POWER_SAVE_MODE_SAMSUNG = "1";
   private static final int POWER_SAVE_MODE_XIAOMI = 1;
   private static final int POWER_SAVE_MODE_HUAWEI = 4;
-
-  /** Plugin registration. */
-  public static void registerWith(PluginRegistry.Registrar registrar) {
-    final BatteryPlusPlugin instance = new BatteryPlusPlugin();
-    instance.onAttachedToEngine(registrar.context(), registrar.messenger());
-  }
 
   @Override
   public void onAttachedToEngine(FlutterPluginBinding binding) {
@@ -59,7 +55,7 @@ public class BatteryPlusPlugin implements MethodCallHandler, StreamHandler, Flut
   }
 
   @Override
-  public void onDetachedFromEngine(FlutterPluginBinding binding) {
+  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
     applicationContext = null;
     methodChannel.setMethodCallHandler(null);
     methodChannel = null;
@@ -68,7 +64,7 @@ public class BatteryPlusPlugin implements MethodCallHandler, StreamHandler, Flut
   }
 
   @Override
-  public void onMethodCall(MethodCall call, Result result) {
+  public void onMethodCall(MethodCall call, @NonNull Result result) {
     if (call.method.equals("getBatteryLevel")) {
       int batteryLevel = getBatteryLevel();
 
@@ -90,13 +86,19 @@ public class BatteryPlusPlugin implements MethodCallHandler, StreamHandler, Flut
     }
   }
 
+  @TargetApi(VERSION_CODES.O)
   @Override
   public void onListen(Object arguments, EventSink events) {
     chargingStateChangeReceiver = createChargingStateChangeReceiver(events);
     applicationContext.registerReceiver(
         chargingStateChangeReceiver, new IntentFilter(Intent.ACTION_BATTERY_CHANGED));
 
-    int status = getBatteryProperty(BatteryManager.BATTERY_PROPERTY_STATUS);
+    int status;
+    if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.LOLLIPOP) {
+      status = getBatteryProperty(BatteryManager.BATTERY_PROPERTY_STATUS);
+    } else {
+      status = BatteryManager.BATTERY_STATUS_UNKNOWN;
+    }
     publishBatteryStatus(events, status);
   }
 
@@ -107,7 +109,7 @@ public class BatteryPlusPlugin implements MethodCallHandler, StreamHandler, Flut
   }
 
   private int getBatteryLevel() {
-    int batteryLevel = -1;
+    int batteryLevel;
     if (VERSION.SDK_INT >= VERSION_CODES.LOLLIPOP) {
       batteryLevel = getBatteryProperty(BatteryManager.BATTERY_PROPERTY_CAPACITY);
     } else {
@@ -170,9 +172,10 @@ public class BatteryPlusPlugin implements MethodCallHandler, StreamHandler, Flut
     return null;
   }
 
+  @RequiresApi(api = VERSION_CODES.LOLLIPOP)
   private int getBatteryProperty(int property) {
     BatteryManager batteryManager =
-        (BatteryManager) applicationContext.getSystemService(applicationContext.BATTERY_SERVICE);
+        (BatteryManager) applicationContext.getSystemService(Context.BATTERY_SERVICE);
     return batteryManager.getIntProperty(property);
   }
 
