@@ -2,5 +2,4 @@ package io.flutter.plugins.batteryexample.example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity: FlutterActivity()
