@@ -1,10 +1,15 @@
 #ifndef NETWORK_MANAGER_H
 #define NETWORK_MANAGER_H
 
+// clang-format off
+#include <winsock2.h>
+// clang-format on
 #include <windows.h>
 
 #include <functional>
 #include <string>
+
+enum class ConnectivityType { None, Ethernet, WiFi };
 
 class NetworkListener;
 struct IConnectionPoint;
@@ -12,7 +17,7 @@ struct IConnectionPointContainer;
 struct INetworkListManager;
 struct IUnknown;
 
-typedef std::function<void(bool)> NetworkCallback;
+typedef std::function<void()> NetworkCallback;
 
 class NetworkManager {
  public:
@@ -22,7 +27,7 @@ class NetworkManager {
   bool Init();
   void Cleanup();
 
-  bool IsConnected();
+  ConnectivityType GetConnectivityType() const;
 
   bool StartListen(NetworkCallback pCallback);
   void StopListen();
@@ -31,6 +36,8 @@ class NetworkManager {
   int GetError() const;
 
  private:
+  std::vector<GUID> GetConnectedAdapterIds() const;
+
   DWORD dwCookie = 0;
   IUnknown *pUnknown = NULL;
   INetworkListManager *pNetworkListManager = NULL;
