@@ -10,6 +10,7 @@ import android.content.pm.FeatureInfo;
 import android.content.pm.PackageManager;
 import android.os.Build;
 import android.provider.Settings;
+import androidx.annotation.NonNull;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import java.util.Arrays;
@@ -35,7 +36,7 @@ class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
   }
 
   @Override
-  public void onMethodCall(MethodCall call, MethodChannel.Result result) {
+  public void onMethodCall(MethodCall call, @NonNull MethodChannel.Result result) {
     if (call.method.equals("getAndroidDeviceInfo")) {
       Map<String, Object> build = new HashMap<>();
       build.put("board", Build.BOARD);
