@@ -6,22 +6,15 @@ package dev.fluttercommunity.plus.network_info;
 
 import android.content.Context;
 import android.net.wifi.WifiManager;
+import androidx.annotation.NonNull;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodChannel;
-import io.flutter.plugin.common.PluginRegistry.Registrar;
 
 /** NetworkInfoPlusPlugin */
 public class NetworkInfoPlusPlugin implements FlutterPlugin {
 
   private MethodChannel methodChannel;
-
-  /** Plugin registration. */
-  public static void registerWith(Registrar registrar) {
-
-    NetworkInfoPlusPlugin plugin = new NetworkInfoPlusPlugin();
-    plugin.setupChannels(registrar.messenger(), registrar.context());
-  }
 
   @Override
   public void onAttachedToEngine(FlutterPluginBinding binding) {
@@ -29,13 +22,14 @@ public class NetworkInfoPlusPlugin implements FlutterPlugin {
   }
 
   @Override
-  public void onDetachedFromEngine(FlutterPluginBinding binding) {
+  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
     teardownChannels();
   }
 
   private void setupChannels(BinaryMessenger messenger, Context context) {
     methodChannel = new MethodChannel(messenger, "dev.fluttercommunity.plus/network_info");
-    WifiManager wifiManager = (WifiManager) context.getSystemService(Context.WIFI_SERVICE);
+    WifiManager wifiManager =
+        (WifiManager) context.getApplicationContext().getSystemService(Context.WIFI_SERVICE);
 
     NetworkInfo networkInfo = new NetworkInfo(wifiManager);
 
