@@ -4,17 +4,17 @@
 
 package dev.fluttercommunity.plus.packageinfo;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.os.Build;
+import androidx.annotation.NonNull;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
-import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
-import io.flutter.plugin.common.PluginRegistry.Registrar;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.HashMap;
@@ -25,43 +25,45 @@ public class PackageInfoPlugin implements MethodCallHandler, FlutterPlugin {
   private Context applicationContext;
   private MethodChannel methodChannel;
 
-  /** Plugin registration. */
-  public static void registerWith(Registrar registrar) {
-    final PackageInfoPlugin instance = new PackageInfoPlugin();
-    instance.onAttachedToEngine(registrar.context(), registrar.messenger());
+  @SuppressWarnings("deprecation")
+  private static long getLongVersionCode(PackageInfo info) {
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+      return info.getLongVersionCode();
+    }
+    return info.versionCode;
   }
 
+  /** Plugin registration. */
   @Override
   public void onAttachedToEngine(FlutterPluginBinding binding) {
-    onAttachedToEngine(binding.getApplicationContext(), binding.getBinaryMessenger());
-  }
-
-  private void onAttachedToEngine(Context applicationContext, BinaryMessenger messenger) {
-    this.applicationContext = applicationContext;
-    methodChannel = new MethodChannel(messenger, "dev.fluttercommunity.plus/package_info");
+    this.applicationContext = binding.getApplicationContext();
+    methodChannel =
+        new MethodChannel(binding.getBinaryMessenger(), "dev.fluttercommunity.plus/package_info");
     methodChannel.setMethodCallHandler(this);
   }
 
   @Override
-  public void onDetachedFromEngine(FlutterPluginBinding binding) {
+  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
     applicationContext = null;
     methodChannel.setMethodCallHandler(null);
     methodChannel = null;
   }
 
   @Override
-  public void onMethodCall(MethodCall call, Result result) {
+  public void onMethodCall(MethodCall call, @NonNull Result result) {
     try {
       if (call.method.equals("getAll")) {
         PackageManager pm = applicationContext.getPackageManager();
         PackageInfo info = pm.getPackageInfo(applicationContext.getPackageName(), 0);
+
+        String buildSignature = getBuildSignature(pm);
 
         Map<String, String> map = new HashMap<>();
         map.put("appName", info.applicationInfo.loadLabel(pm).toString());
         map.put("packageName", applicationContext.getPackageName());
         map.put("version", info.versionName);
         map.put("buildNumber", String.valueOf(getLongVersionCode(info)));
-        map.put("buildSignature", getBuildSignature(pm));
+        if (buildSignature != null) map.put("buildSignature", buildSignature);
 
         result.success(map);
       } else {
@@ -73,24 +75,33 @@ public class PackageInfoPlugin implements MethodCallHandler, FlutterPlugin {
   }
 
   @SuppressWarnings("deprecation")
-  private static long getLongVersionCode(PackageInfo info) {
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
-      return info.getLongVersionCode();
-    }
-    return info.versionCode;
-  }
-
   private String getBuildSignature(PackageManager pm) {
     try {
-      PackageInfo packageInfo =
-          pm.getPackageInfo(applicationContext.getPackageName(), PackageManager.GET_SIGNATURES);
-      if (packageInfo == null
-          || packageInfo.signatures == null
-          || packageInfo.signatures.length == 0
-          || packageInfo.signatures[0] == null) {
-        return null;
+      if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+        PackageInfo packageInfo =
+            pm.getPackageInfo(
+                applicationContext.getPackageName(), PackageManager.GET_SIGNING_CERTIFICATES);
+        if (packageInfo == null || packageInfo.signingInfo == null) {
+          return null;
+        }
+        if (packageInfo.signingInfo.hasMultipleSigners()) {
+          return signatureToSha1(packageInfo.signingInfo.getApkContentsSigners()[0].toByteArray());
+        } else {
+          return signatureToSha1(
+              packageInfo.signingInfo.getSigningCertificateHistory()[0].toByteArray());
+        }
+      } else {
+        @SuppressLint("PackageManagerGetSignatures")
+        PackageInfo packageInfo =
+            pm.getPackageInfo(applicationContext.getPackageName(), PackageManager.GET_SIGNATURES);
+        if (packageInfo == null
+            || packageInfo.signatures == null
+            || packageInfo.signatures.length == 0
+            || packageInfo.signatures[0] == null) {
+          return null;
+        }
+        return signatureToSha1(packageInfo.signatures[0].toByteArray());
       }
-      return signatureToSha1(packageInfo.signatures[0].toByteArray());
     } catch (PackageManager.NameNotFoundException | NoSuchAlgorithmException e) {
       return null;
     }
