@@ -11,7 +11,8 @@ import 'package:flutter/material.dart';
 import 'package:sensors_plus/sensors_plus.dart';
 
 class Snake extends StatefulWidget {
-  Snake({this.rows = 20, this.columns = 20, this.cellSize = 10.0}) {
+  Snake({Key? key, this.rows = 20, this.columns = 20, this.cellSize = 10.0})
+      : super(key: key) {
     assert(10 <= rows);
     assert(10 <= columns);
     assert(5.0 <= cellSize);
@@ -22,6 +23,7 @@ class Snake extends StatefulWidget {
   final double cellSize;
 
   @override
+  // ignore: no_logic_in_create_state
   State<StatefulWidget> createState() => SnakeState(rows, columns, cellSize);
 }
 
