@@ -1,4 +1,5 @@
 import 'dart:html' as html;
+import 'dart:developer' as developer;
 
 /// Receive permission status of the API.
 Future<void> checkPremission(
@@ -20,21 +21,21 @@ Future<void> checkPremission(
         initSensor();
       } else if (premissionStatus.state == 'prompt') {
         /// user needs to intract with this
-        print(
+        developer.log(
           'Premission [$premissionName] still has not been granted or denied.',
         );
       } else {
         // If permission is denied, do not do anything
-        print('Permission [$premissionName] to use sensor was denied.');
+        developer.log('Permission [$premissionName] to use sensor was denied.');
       }
     } catch (e) {
-      print(
+      developer.log(
         'Integration with Permissions API is not enabled, still try to start app.',
       );
       initSensor();
     }
   } else {
-    print('No Permissions API, still try to start app.');
+    developer.log('No Permissions API, still try to start app.');
     initSensor();
   }
 }
