@@ -210,7 +210,9 @@ static UIViewController *TopViewControllerForViewController(UIViewController *vi
            subject:(NSString *)subject
     withController:(UIViewController *)controller
           atSource:(CGRect)origin {
-  SharePlusData *data = [[SharePlusData alloc] initWithSubject:subject text:shareText];
+  NSObject *data = [shareText hasPrefix:@"https://"]
+                       ? [[NSURL alloc] initWithString:shareText]
+                       : [[SharePlusData alloc] initWithSubject:subject text:shareText];
   [self share:@[ data ] withController:controller atSource:origin];
 }
 
